// apps/web/src/contracts/generated/FeatureChange.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Entity } from "./Entity";

export type FeatureChange = { "op": "create", id: string, entity: Entity, } | { "op": "update", id: string, entity: Entity, } | { "op": "delete", id: string, };